refactor: styles 파일 분리
