+++ src/components/shared/Modal/index.tsx
@@ -1,38 +1,10 @@
 import { FC, MouseEvent } from 'react';
-import styled, { css } from 'styled-components';
-import PALETTE from '../../../constants/palette';
+import { ModalContainer, ModalInner } from './styles';
 
-interface Props {
+export interface Props {
   onClose?: () => void;
   type?: 'bottom' | 'full';
 }
-
-const ModalContainer = styled.div`
-  display: flex;
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background: ${PALETTE.TRANSLUCENT_BLACK_9};
-`;
-
-const bottomType = css`
-  margin-top: auto;
-  height: 16rem;
-  border-radius: 1rem 1rem 0 0;
-`;
-
-const fullType = css`
-  height: 100vh;
-`;
-
-const ModalInner = styled.div<Pick<Props, 'type'>>`
-  position: relative;
-  width: 100%;
-  background: white;
-  ${({ type }) => (type === 'full' ? fullType : bottomType)};
-`;
 
 const Modal: FC<Props> = ({ children, onClose, type }) => {
   const onClickDimmed = ({ target, currentTarget }: MouseEvent<HTMLDivElement>) => {
