+++ src/components/shared/Template/styles.ts
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const PageContainer = styled.div`
+  padding: 1rem 1.125rem;
+`;
+
+export const Header = styled.header`
+  display: flex;
+  align-items: center;
+
+  a {
+    display: flex;
+    align-items: center;
+    margin-right: 0.25rem;
+  }
+`;
